patterns/03-multi-tenant-keys/backend/pyproject.toml
[project]
name = "multi-tenant-keys"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "genkit",
    "genkit-plugin-google-genai",
    "genkit-plugin-fastapi",
    "fastapi",
    "uvicorn",
    "firebase-admin",
    "google-cloud-secret-manager",
]
